Fix a OapStrategy Issue.
OapSortLimitStrategy should support support deserialized logicalPlan.
Normally we call sql like sql("").collect which uses plain logicalPlan,
so that OapSortLimitStrategy is applicable.
But sometimes it can be written as: sql("").foreach{...} which will
refer to a deserialized logicalPlan as output of each rdd should be
deserialized. OapSortLimitStrategy should also work with this case.

Signed-off-by: Wang, Yue A <yue.a.wang@intel.com>